app/requirements.txt
click==8.1.3
Flask==2.1.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
typing-extensions==4.3.0
Werkzeug==2.1.2
zipp==3.8.1
mysql-connector-python==8.0.29
